LayoutTests/fast/dom/Window/dispatchEvent-expected.txt

Test window.dispatchEvent().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.dispatchEvent(event) threw exception Error: UNSPECIFIED_EVENT_TYPE_ERR: DOM Events Exception 0.
PASS window.dispatchEvent(event) threw exception Error: UNSPECIFIED_EVENT_TYPE_ERR: DOM Events Exception 0.
PASS myEventDispatched is true
PASS target is window
PASS currentTarget is window
PASS useCaptureDispatched is true
PASS nonUseCaptureDispatched is true
PASS successfullyParsed is true

TEST COMPLETE
